requirements.txt
asgiref==3.7.2
Brotli==1.1.0
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
gunicorn==21.2.0
packaging==23.2
platformdirs==4.1.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4
virtualenv==20.25.0
whitenoise==6.6.0
